backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
pydantic==2.11.4
pydantic-settings==2.9.1
python-multipart==0.0.20
websockets==15.0.1
redis==5.3.0
celery==5.5.3
httpx==0.28.1

# RAG (Pinecone + LangChain-compatible interfaces)
pinecone==7.3.0

# Gmail API + OAuth2
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2

# Gemini
google-genai


anthropic==0.42.0

# Groq (Llama 3)
groq
